Stop tab navigation entirely
I think I basically need to re-implement this myself, unfortunately. It picks
up a bunch of extra stuff I don't want to be able to tab to, and it loses
position when switching tabs. Part of that might be because I'm creating the
editors newly every frame (instead of just calling show???)

I think I am going to play around with that part more and see if I can minimize
the code that I actually have to write