Add a garbage collector to the TextBoxStore
This is probably unnecessary, there's not going to be that many
discarded textboxes over the lifetime of the application. However, it's
still good to have, and easy to implement. Also it's a very funny way to
do a garbage collector.

Unrelatedly, I am begging Rust to make the syntax for quickly generating
a random number less tedious.