Create TextBoxContext and store it
A data structure that persists from one frame to the next and stores all
the info we want to preserve for a textbox. We take advantage of the
fact that it's always correct to rebuild this if needed and do some
slightly funky things (using the memory address of the string as an
identifier).