Fix bug with top level folders being renamed
I'm not totally sure how this bug was happening in the first place: it
would have required the top level folders to get the modified state
somehow, which shouldn't have been happening, but they shouldn't be
renamed anymore, which is good

This also sets the name (in metadata) of the top level folders to be
"Text" instead of "text". I'm not sure if this was totally necessary, it
feels a bit silly to write extra code to handle that case, but I think
it looks visually nicer. This probably won't cause any new bugs and
future Brie won't be upset about this