Fix crash in fix indexing (#148) (#129)
It turns out that I was just being silly and not understanding when
borrows actually get dropped. I was thinking that the extra function
call would result in the borrow being dropped already, but that isn't
the case. I think I had another bug with almost the same issue
recently...

At least I was *slightly* less confused this time