Refactor Code to use Schemas
A new FileType type (well technically it's an alias) is created. This
will allow the code to access properties of a file type without needing
an instance of an object of that type.

The code is now fully agnostic to the content of FileObject. They are
created through a schema, and the schema is the only part of the code
that can see the FileObjects that exist. The rest of the code only ever
interfaces with them through the FileObject and FIleObjectEditor trait.

This commit lays the groundwork for allowing different cheese-paper
projects to use different schemas, but does not implemnt that. For now,
all projects use the default schema.