Recursively rescan indexing when loading folders
If a folder was modified in place, we would reload the file object,
which wouldn't update indexing. This fixes that bug

Eve might tell me that I'm chasing down endless bugs here still, but I
think this is a required change, we still need to reload the file
contents (to not overwrite with old versions of in-editor changes when
they get saved), it's just a scenario that I didn't think about.

It should hopefully fix the tests being broken on windows, where we see
test_tracker_new_file_index getting broken by the text file also
receiving a modify (which was how I discovered this issue). It might
also be good to add more checks that we don't have any files without
indexes aside from the top level folders